estadisticas

+++ Code/Fork/ProcessPool.h
@@ -9,8 +9,9 @@ typedef int (*ProcessTask)(size_t index, void *context, void *result);
 
 /* Ejecuta `task` para los índices 0..count-1, cada uno en un proceso hijo
  * (a lo sumo uno por núcleo a la vez). El resultado del hijo i queda en
- * results + i * result_size. `results` puede ser NULL si result_size es 0. */
-int process_pool_run(size_t count, ProcessTask task, void *context,
-                     void *results, size_t result_size);
+ * results + i * result_size. `results` puede ser NULL si result_size es 0.
+ * Devuelve cuántos hijos terminaron bien (count si no falló ninguno). */
+size_t process_pool_run(size_t count, ProcessTask task, void *context,
+                        void *results, size_t result_size);
 
 #endif
