+++ Code/Pthread/ThreadPool.h
@@ -11,7 +11,8 @@ void *shared_memory_create(size_t size);
 void shared_memory_destroy(void *memory, size_t size);
 
 /* Ejecuta `task` para los índices 0..count-1 repartidos entre varios hilos
- * (uno por núcleo) que toman trabajo de una cola en memoria compartida. */
-int thread_pool_run(size_t count, ThreadTask task, void *context);
+ * (uno por núcleo) que toman trabajo de una cola en memoria compartida.
+ * Devuelve cuántos elementos salieron bien (count si no falló ninguno). */
+size_t thread_pool_run(size_t count, ThreadTask task, void *context);
 
 #endif
